Move motorResources to pkg/client - there were duplicate GRPC connections
