+++ pkg/motor/create_schema.go
@@ -37,7 +37,7 @@ func (mtr *motorNodeImpl) CreateSchema(request mt.CreateSchemaRequest) (mt.Creat
 	}
 
 	// store reference to newly created WhatIs
-	_, err = mtr.Resources.StoreWhatIs(csresp.WhatIs)
+	_, err = mtr.GetClient().StoreWhatIs(csresp.WhatIs)
 	if err != nil {
 		return mt.CreateSchemaResponse{}, fmt.Errorf("store WhatIs: %s", err)
 	}
