+++ pkg/motor/object_builder.go
@@ -10,7 +10,7 @@ import (
 )
 
 func (mtr *motorNodeImpl) NewObjectBuilder(did string) (*object.ObjectBuilder, error) {
-	whatIs, schema, found := mtr.Resources.GetSchema(did)
+	whatIs, schema, found := mtr.GetClient().GetSchema(did)
 	if !found {
 		return nil, fmt.Errorf("could not find WhatIs with did '%s'", did)
 	}
