+++ pkg/motor/object_builder_test.go
@@ -1,7 +1,6 @@
 package motor
 
 import (
-	"context"
 	"fmt"
 	"testing"
 
@@ -11,14 +10,14 @@ import (
 )
 
 /*
-prt.CreateSchemaRequest{
-	Label: "TestUser",
-	Fields: map[string]prt.CreateSchemaRequest_SchemaKind{
-		"email":     prt.CreateSchemaRequest_SCHEMA_KIND_STRING,
-		"firstName": prt.CreateSchemaRequest_SCHEMA_KIND_STRING,
-		"age":       prt.CreateSchemaRequest_SCHEMA_KIND_INT,
-	},
-}
+	prt.CreateSchemaRequest{
+		Label: "TestUser",
+		Fields: map[string]prt.CreateSchemaRequest_SchemaKind{
+			"email":     prt.CreateSchemaRequest_SCHEMA_KIND_STRING,
+			"firstName": prt.CreateSchemaRequest_SCHEMA_KIND_STRING,
+			"age":       prt.CreateSchemaRequest_SCHEMA_KIND_INT,
+		},
+	}
 */
 const SCHEMA_DID string = "did:snr:QmZLKGrTcUAKsUVUZ5e72rAWRg1Y1SzRJqWqcXaDqjFUqm"
 
@@ -43,10 +42,7 @@ func Test_ObjectBuilder(t *testing.T) {
 		assert.NoError(t, err, "login succeeds")
 
 		// query WhatIs so it's cached
-		_, err = m.QueryWhatIs(context.Background(), mt.QueryWhatIsRequest{
-			Creator: m.GetDID().String(),
-			Did:     SCHEMA_DID,
-		})
+		_, err = m.GetClient().QueryWhatIs(m.GetDID().String(), SCHEMA_DID)
 		assert.NoError(t, err, "query whatis")
 
 		// upload object
